+++ src/components/experiment-builder/experiment-builder.jsx
@@ -1,12 +1,12 @@
 import "./experiment-builder.css"
 
-import NewFlexibleFormField from "../../UI/FormComponents/NewFlexibleFormField"
+import FlexibleFormField from "../ui/flexible-form-field/flexible-form-field"
 import React from "react"
 
 import { Button, ButtonGroup, Typography } from "@mui/material"
 import { capitalizeFirstLetter } from "../../utils/string-helpers"
 import { Field, Form, Formik } from "formik"
-import { SmartFolderField } from "../../UI/FormComponents/FieldTypes/SmartFolderField"
+import { FolderField } from "../ui/folder-field"
 
 const nestedFolders = [
   "models",
@@ -49,7 +49,7 @@ const ExperimentBuilder = ({ FormikProps, handleFileChange, newQuestions, setFie
           {/* Case 1: Folder is nested. */}
           { nestedFolders.includes(folder) ? (
             <>
-              <SmartFolderField folder = { folder } folderContent = { folderContent } />
+              <FolderField folder = { folder } folderContent = { folderContent } />
 
               {/* Loop each file in each folder. */}
               { Object.entries(folderContent).map(([file, fileContent]) => (
@@ -81,7 +81,7 @@ const ExperimentBuilder = ({ FormikProps, handleFileChange, newQuestions, setFie
                                         { param }:
                                       </label>
                                       
-                                      <NewFlexibleFormField
+                                      <FlexibleFormField
                                         id = { `${folder}_params:${className}:${param}` }
                                         object = { value }
                                         type = { value.type }
@@ -119,7 +119,7 @@ const ExperimentBuilder = ({ FormikProps, handleFileChange, newQuestions, setFie
                   
                   <label htmlFor = { `${folder}:${param}` }>{param}:</label>
                   
-                  <NewFlexibleFormField
+                  <FlexibleFormField
                     id = { `${folder}:${param}` }
                     object = { value }
                     type = { value.type }
@@ -135,7 +135,7 @@ const ExperimentBuilder = ({ FormikProps, handleFileChange, newQuestions, setFie
               
               <label htmlFor = { `${folder}` }>{ folder }:</label>
               
-              <NewFlexibleFormField
+              <FlexibleFormField
                 id = { `${folder}` }
                 object = { folderContent }
                 type = { folderContent.type }
