+++ src/components/ui/flexible-form-field/flexible-form-field.jsx
@@ -1,13 +1,13 @@
-import "./Form.css"
+import "./flexible-form-field.css"
 
-import BooleanField from "../../components/ui/boolean-field"
-import DefaultTextField from "../../components/ui/default-text-field"
-import FormulaField from "../../components/ui/formula-field"
+import BooleanField from "../boolean-field.jsx"
+import DefaultTextField from "../default-text-field.jsx"
+import FormulaField from "../formula-field.jsx"
 import React from "react"
 
-import FloatField from "../../components/ui/float-field.jsx"
-import StringField from "../../components/ui/string-field.jsx"
-import IntegerField from "../../components/ui/integer-field.jsx"
+import FloatField from "../float-field.jsx"
+import StringField from "../string-field.jsx"
+import IntegerField from "../integer-field.jsx"
 
 
 import { useFormikContext } from "formik"
